pyproject.toml
[tool.poetry]
name = "clustering"
version = "0.2.1"
description = "WM atlas"
authors = ["Edouard Koehn"]
license = "MIT License"

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
pyyaml = "^6.0"
mat73 = "^0.60"
scipy = "^1.10.1"
networkx = {extras = ["default"], version = "^3.0"}
matplotlib = "^3.7.1"
seaborn = "^0.12.2"
scikit-learn = "^1.2.2"
click = "^8.1.3"
pandas = "^1.5.3"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "^21.12b0"
pre-commit = "^2.17.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
exclude = '''
/(
    \.eggs
  | \.git
  | \.venv
  | \.env
  | build
  | dist
  | \.mypy_cache
)\
'''

[tool.poetry.scripts]
spectral_clustering = "clustering:spectral_clustering.spectral_clustering"

[tool.commitizen]
version = "0.2.1"
tag_format = "v$version"
version_files = [
  "pyproject.toml:version",
  "application/__version__.py"
]
